# docs/swagger.yaml
basePath: /api
definitions:
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.UploadFileResponse:
    properties:
      id:
        type: string
      url:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  title: URFU-Guide
  version: "1.0"
paths:
...
  /file/upload:
    post:
      consumes:
      - image/jpeg
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UploadFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Upload file
      tags:
      - file
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Token
    type: apiKey
swagger: "2.0"